웹코드/requirements.txt
absl-py==2.3.0
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
grpcio==1.72.1
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
path==17.1.0
pathlib==1.0.1
Pil-Lite==0.1.3
pillow==11.2.1
protobuf==6.31.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2024.8.30
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
tqdm_multiprocessing==0.1.2
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.22.0
